% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.supports_unicode}
\alias{.supports_unicode}
\title{Safe Unicode Display Functions}
\usage{
.supports_unicode()
}
\description{
Functions to safely display Unicode characters with fallbacks
for systems that don't support them properly.
}
