% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCriticalValue.R
\name{getCriticalValue}
\alias{getCriticalValue}
\title{Identify the critical value for the exact risk difference test}
\usage{
getCriticalValue(
  alpha,
  piTreatment,
  piControl,
  nTreatment,
  nControl,
  mustExceed = TRUE
)
}
\arguments{
\item{alpha}{One-sided significance level.}

\item{piTreatment}{Event rate in treatment group.}

\item{piControl}{Event rate in control group.}

\item{nTreatment}{Number of subjects in treatment group.}

\item{nControl}{Number of subjects in control group.}

\item{mustExceed}{Logical indicating whether the observed risk difference must exceed the critical value for rejection (default \code{TRUE}). IF \code{FALSE}, the observed risk difference must be equal to or greater than the critical value.}
}
\value{
Critical value.
}
\description{
Identify the critical value for the exact risk difference test
}
