% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPower.R
\name{getPower}
\alias{getPower}
\title{Calculate power for given success rates}
\usage{
getPower(
  piTreatment,
  piControl,
  nTreatment,
  nControl,
  alpha,
  piNull = c(piControl, piControl)
)
}
\arguments{
\item{piTreatment}{Treatment success rate.}

\item{piControl}{Control success rate.}

\item{nTreatment}{Number of subjects in treatment group.}

\item{nControl}{Number of subjects in control group.}

\item{alpha}{One-sided significance level.}

\item{piNull}{Null hypothesis success rates. Should consist of two entries, the first for the treatment success rate under the null hypothesis, the second for the control success rate under the null hypothesis. Defaults to \code{piControl} for both entries.}
}
\value{
The power for the given scenario.
}
\description{
Calculate power for given success rates
}
