% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotriskdiff.R
\name{plotriskdiff}
\alias{plotriskdiff}
\title{Plot risk difference}
\usage{
plotriskdiff(agg, title = "")
}
\arguments{
\item{agg}{Data frame with columns rds and x, created by \code{driskdiff} or \code{priskdiff}.}

\item{title}{Optional plot title.}
}
\description{
Plot risk difference
}
