% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iptw_causal.R
\name{print.iptw_result}
\alias{print.iptw_result}
\title{Print Method for IPTW Results}
\usage{
\method{print}{iptw_result}(x, ...)
}
\arguments{
\item{x}{An iptw_result object}

\item{...}{Additional arguments passed to print}
}
\description{
Print Method for IPTW Results
}
