% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iptw_causal.R
\name{print.nnt_iptw_result}
\alias{print.nnt_iptw_result}
\title{Print Method for IPTW NNT Results}
\usage{
\method{print}{nnt_iptw_result}(x, digits = 1, ...)
}
\arguments{
\item{x}{An nnt_iptw_result object from calc_risk_diff_iptw(..., nnt = TRUE)}

\item{digits}{Number of decimal places for NNT estimates (default: 1)}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for IPTW NNT Results
}
