% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{print.nnt_result}
\alias{print.nnt_result}
\title{Print Method for NNT Results}
\usage{
\method{print}{nnt_result}(x, digits = 1, ...)
}
\arguments{
\item{x}{An nnt_result object from calc_risk_diff(..., nnt = TRUE)}

\item{digits}{Number of decimal places for NNT estimates (default: 1)}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for NNT Results
}
