% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_risk_diff.R
\name{print.riskdiff_result}
\alias{print.riskdiff_result}
\title{Print method for riskdiff_result objects}
\usage{
\method{print}{riskdiff_result}(x, ...)
}
\arguments{
\item{x}{A riskdiff_result object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for riskdiff_result objects
}
