% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priskdiff.R
\name{priskdiff}
\alias{priskdiff}
\title{Distribution function of exact risk difference}
\usage{
priskdiff(piTreatment, piControl, nTreatment, nControl)
}
\arguments{
\item{piTreatment}{Event rate in treatment group.}

\item{piControl}{Event rate in control group.}

\item{nTreatment}{Number of subjects in treatment group.}

\item{nControl}{Number of subjects in control group.}
}
\value{
Aggregated data of risk differences and cumulative probabilities.
}
\description{
Distribution function of exact risk difference
}
